+++ config/initializers/column_definition.rb
@@ -15,7 +15,11 @@
 # Read more at https://github.com/publiclab/mapknitter/pull/323
 
 class ActiveRecord::ConnectionAdapters::ColumnDefinition
-  def sql_type
-    type.to_sym == :primary_key ? 'int(11) auto_increment PRIMARY KEY' : base.type_to_sql(type.to_sym, limit, precision, scale) rescue type
+
+  if ActiveRecord::Base.connection.adapter_name != 'sqlite3' && ActiveRecord::Base.connection.adapter_name != 'SQLite'
+    def sql_type
+      type.to_sym == :primary_key ? 'int(11) auto_increment PRIMARY KEY' : base.type_to_sql(type.to_sym, limit, precision, scale) rescue type
+    end 
   end 
+
 end
