+++ start.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+sleep 5
+bundle exec rails s -p 3000 -b '0.0.0.0'
