+++ test/unit/warpable_test.rb
@@ -22,17 +22,14 @@ class WarpableTest < ActiveSupport::TestCase
   end
 
   test "try export" do
+    # make a sample image
     system('mkdir -p public/system/images/1/original')
     system('cp test/fixtures/demo.png public/system/images/1/original/')
     system('mkdir -p public/warps/saugus-landfill-incinerator')
     system('touch public/warps/saugus-landfill-incinerator/folder')
     assert File.exist?('public/warps/saugus-landfill-incinerator/folder')
-    w = warpables(:one)    
+    w = warpables(:one)
     assert_not_nil w.save_dimensions
-    assert_not_nil w.generate_perspectival_distort(10, w.map.slug)
-    assert_not_nil w.delete_temp_files(w.map.slug)
-    assert_not_nil w.working_directory(w.map.slug)
-    assert_not_nil w.warps_directory(w.map.slug)
     assert_not_nil w.user_id 
     assert File.exist?('public/warps/saugus-landfill-incinerator/1-geo.tif')
   end
